+++ Baitap/bai17/quanlythuvien/categories/delete.php
@@ -7,9 +7,14 @@ include_once '../database.php';
 // die();
 //
 $id = $_REQUEST['id'];
+if (isset($_REQUEST['id'])){
+
 // echo $id;
 $sql = "DELETE FROM `categories` WHERE `id` = $id ";
 // echo $sql;
 $conn->exec($sql);
 //chuyen huong ve trang list.php
 header('location:index.php');
+} else {
+    header('location:../err.php');
+}
