+++ Baitap/bai17/quanlythuvien/student/delete.php
@@ -6,10 +6,17 @@ include_once '../database.php';
 // print_r($_REQUEST);
 // die();
 //
-$id = $_REQUEST['id'];
-// echo $id;
-$sql = "DELETE FROM `students` WHERE `id` = $id ";
-// echo $sql;
-$conn->exec($sql);
-//chuyen huong ve trang list.php
-header('location:index.php');
+$id = $_REQUEST['id']; 
+if (isset($_REQUEST['id'])){
+
+    // echo $id;
+    $sql = "DELETE FROM `students` WHERE `id` = $id ";
+    // echo $sql;
+    $conn->exec($sql);
+    //chuyen huong ve trang list.php
+    header('location:index.php');
+}
+else{
+    header('location:../err.php');
+
+}
